Refactor out functions in zip_utils to tmpfile_utils
