+++ test/test_misc_utils.py
@@ -31,12 +31,14 @@ from dcicutils.misc_utils import (
     ObsoleteError, CycleError, TopologicalSorter, keys_and_values_to_dict, dict_to_keys_and_values, is_c4_arn,
     deduplicate_list, chunked, parse_in_radix, format_in_radix, managed_property, future_datetime,
     MIN_DATETIME, MIN_DATETIME_UTC, INPUT, builtin_print, map_chunked, to_camel_case, json_file_contents,
-    pad_to, JsonLinesReader, split_string, merge_objects, to_integer
+    pad_to, JsonLinesReader, split_string, merge_objects, to_integer,
+    load_json_from_file_expanding_environment_variables
 )
 from dcicutils.qa_utils import (
     Occasionally, ControlledTime, override_environ as qa_override_environ, MockFileSystem, printed_output,
     raises_regexp, MockId, MockLog, input_series,
 )
+from dcicutils.tmpfile_utils import temporary_file
 from typing import Any, Dict, List
 from unittest import mock
 
@@ -3691,3 +3693,11 @@ def test_to_integer():
     assert to_integer("0.0") == 0
     assert to_integer("asdf") is None
     assert to_integer("asdf", "myfallback") == "myfallback"
+
+
+def test_load_json_from_file_expanding_environment_variables():
+    with mock.patch.object(os, "environ", {"Auth0Secret": "dgakjhdgretqobv", "SomeEnvVar": "xyzzy"}):
+        some_json = {"Auth0Secret": "${Auth0Secret}", "abc": "def", "someproperty": "$SomeEnvVar"}
+        with temporary_file(content=json.dumps(some_json), suffix=".json") as tmpfile:
+            expanded_json = load_json_from_file_expanding_environment_variables(tmpfile)
+            assert expanded_json == {"Auth0Secret": "dgakjhdgretqobv", "abc": "def", "someproperty": "xyzzy"}
